backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlmodel==0.0.24
pydantic-settings==2.8.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
psycopg[binary]==3.2.6
